# public/openapi.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "https://syncal.api.teamstep.io/v1/"
        }
    ],
    "info": {
        "description": "API references for the collaborative calendar application, SynCal",
        "version": "1.0.0",
        "title": "SynCal API Reference",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "dev@teamstep.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "All API calls related to the user object"
        }
    ],
    "paths": {
        "/user/:firstName": {
            "get": {
                "tags": ["User"],
                "summary": "Returns a list of users by the given name",
                "description": "",
                "operationId": "getUsersByName",
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "path",
                        "description": "The first name of the user that should be returned",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/new-user": {
            "post": {
                "tags": ["User"],
                "summary": "Adds a new user to the server with the given information in the body",
                "description": "",
                "parameters": [],
                "operationId": "createNewUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "format": "int32"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "description": "The user information that you want to add"
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "age": {
                        "type": "number",
                        "format": "int32"
                    }
                }
            }
        },
        "links": {},
        "callbacks": {}
    },
    "security": []
}